refactor(core/layout): phase 2-3 — relocate flat-tree algorithm
Phase 2: flat-tree/ → auto-placement/flat-tree/

  Pure relocation. Internal imports adjusted (../../models → ../../../models, ../tree-layout → ../../tree-layout). flat-tree-layout.ts shim updated to point at the new path. layoutNetwork still imports through the shim, so no external surface moved yet.

Phase 3: tree-layout.ts → auto-placement/flat-tree/tidy-tree.ts

  Per the design doc, Buchheim is an algorithm primitive, not a rule. It now lives privately inside the flat-tree algorithm directory. The auto-placement entry imports it as ./tidy-tree.js; nothing outside the algorithm references it. (Promote to a shared primitives/ location later if a second auto-placement algorithm wants it.)

Tests unchanged: 171 / 171 still pass. No behavioural change — just file moves with import-path adjustments.

Next phases (this same branch / PR):
  4. Move port-placement.ts into auto-placement/flat-tree/ and split into port-side / port-position. Update decidePortSides to emit the richer PortsBySide (PortInfo[] per side) the engine expects.
  5. Build autoLayoutFlatTree(graph, engine, options) — the new entry that uses engine queries instead of inline arithmetic.
  6. Update consumers, delete legacy network-layout / port-placement / measure-text / flat-tree-layout / unified-engine.
