+++ libs/@shumoku/core/src/layout/auto-placement/flat-tree/engine.api.test.ts
@@ -1,7 +1,7 @@
 // Public facade API.
 
 import { describe, expect, test } from 'vitest'
-import type { Link, NetworkGraph, Node, Subgraph } from '../../models/types.js'
+import type { Link, NetworkGraph, Node, Subgraph } from '../../../models/types.js'
 import { createFlatTreeEngine } from './engine.js'
 
 function node(id: string, parent?: string): Node {
