+++ libs/@shumoku/core/src/layout/auto-placement/flat-tree/internal.ts
